nlsocket: fake `fileno()` fo async mode
The issue with the async mode is that it is not usable for
polling — the `poll()` results on a socket with async cache
are nor reliable, nor even relevant sometimes.

So create a pipe that will mimic I/O for the polling — the
async cache pushes a byte on every received message to the
pipe and it triggers `poll()`.

To avoid pipe overflow, all bytes must be consumed. It does
the `get()` method, which is called from the main thread. It
reads one byte for every message.

There is an issue with the netlink proxy: it may push
messages to the nlsocket's backlog. Thus it must also notify
`poll()` by writing to the pipe.

Normally, the `fileno()` method returns socket's fd. But for
the async mode it returns the pipe's fd.